+++ Activities/GameActivity.cpp
@@ -1491,7 +1491,7 @@ void GameActivity::Update()
 
             // Find the actor closest to the cursor, if any within the radius
 			Vector markedDistance;
-            Actor *pMarkedActor = g_MovableMan.GetClosestTeamActor(team, player, m_ActorCursor[player], g_SceneMan.GetSceneWidth(), markedDistance);
+            Actor *pMarkedActor = g_MovableMan.GetClosestTeamActor(team, player, m_ActorCursor[player], g_SceneMan.GetSceneWidth(), markedDistance, true);
 //            Actor *pMarkedActor = g_MovableMan.GetClosestTeamActor(team, player, m_ActorCursor[player], g_FrameMan.GetPlayerScreenWidth() / 4);
 
             // Player canceled selection of actor